Added Shuttle code and GitHub deploy action.
