+++ .github/workflows/shuttle-deploy.yml
@@ -1,0 +1,15 @@
+name: Shuttle deploy
+
+on:
+  push:
+    branches:
+      - "main"
+      - "heroku_deploy"
+
+jobs:
+  deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: shuttle-hq/deploy-action@v0.0.3
+        with:
+          deploy-key: ${{ secrets.SHUTTLE_DEPLOY_KEY }}
