+++ api/Cargo.toml
@@ -3,9 +3,12 @@ name = "calandar-api"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+shuttle-service = { version = "0.5.2", features = ["web-rocket", "sqlx-postgres"] }
 rocket = "0.5.0-rc.2"
-sqlx = { version = "0.6.2", features = [ "runtime-tokio-rustls", "postgres", "offline" ] }
+sqlx = { version = "0.6.2", features = [ "runtime-tokio-native-tls", "postgres", "offline" ] }
 tokio = { version = "1.21.1", features = ["full"] }
